BUG: Only copy the ITK .clang-format file if the file does not exist
To address:

   Ref 'HEAD' was rewritten
   error: Untracked working tree file '.clang-format' would be overwritten by merge.

+++ entrypoint.sh
@@ -6,7 +6,9 @@ error_message="$1"
 
 cd "$GITHUB_WORKSPACE"
 
-cp /ITK.clang-format ./.clang-format
+if ! test -f ./.clang-format; then
+  cp /ITK.clang-format ./.clang-format
+fi
 /clang-format.bash --tracked
 if ! git diff-index --diff-filter=M --quiet HEAD -- ':!.clang-format'; then
   echo "::error ::${error_message}"
